new `lt-merge` command to merge LU's from BEG to END tag
Since we need to unquote when generating before tf-inject, we need to
double-quote escaped chars here:

  $ echo '^ikke/ikke<adv>$ ^«/«<lquot><MERGE_BEG>$^til/til<pr>$ ^x\@y.com/x\@y.com<email>$^»/»<rquot><MERGE_END>$ ^da/da<adv>$' | lttoolbox/lt-merge
  ^ikke/ikke<adv>$ ^«til x\\\@y.com»/«til x\\\@y.com»<MERGED>$ ^da/da<adv>$

  $ echo '^«/«<lquot><MERGE_BEG>$[[tf:i:a]]^veldig/veldig<adv>$[[/]]^»/»<rquot><MERGE_END>$' | lttoolbox/lt-merge
  ^«\[\[tf:i:a\]\]veldig\[\[\/\]\]»/«\[\[tf:i:a\]\]veldig\[\[\/\]\]»<MERGED>$

If we run this between analysis and wblank-attach, then after
the `lt-proc -b generator.bin` step we should have e.g.

  ^ikkje<adv>/ikkje$ ^«til x\\\@y.com»<MERGED>/«til x\\\@y.com»$ ^då<adv>/då$

which after `cg-proc -1 -n -g genprefs.bin` would turn into

  ikkje «til x\@y.com» då

Note how \\\@ turned into \@ – we removed one layer of quoting, but
this is still in the apertium stream so special chars stay quoted
until the final tf-inject.

TODO:

We need to be able to pass MERGED stuff unchanged through biltrans and
generator, would like to <re>.+</re><i><s n="MERGED"/></i> but . is
literal period in re(!) and even ANY_CHAR doesn't seem supported in
lt-proc -b. It should be possible to support with a `step_case_override`
in `FSTProcessor::biltrans`.

We need an `lt-merge --unmerge` to undo the merge:

  $ echo '^ikkje<adv>/ikkje$ ^«\[\[tf:i:a\]\]s\\\^å\[\[\/\]\]»<MERGED>/«\[\[tf:i:a\]\]s\\\^å\[\[\/\]\]»$' | lt-merge --unmerge
  ^ikkje<adv>/ikkje$ «[[tf:i:a]]s\^å[[/]]»

which then becomes

  $ echo '^ikkje<adv>/ikkje$ «[[tf:i:a]]s\^å[[/]]»' |cg-proc -1ng nob-nno.genprefs.rlx.bin
  ikkje «[[tf:i:a]]s\^å[[/]]»

which tf-inject is happy to handle.